Move stuff to a more common place to reuse between auth providers
