+++ packages/auth/src/getCurrentUserFromMiddleware.ts
