+++ packages/auth/src/index.ts
@@ -2,7 +2,10 @@ export type { AuthContextInterface, CurrentUser } from './AuthContext.js'
 export { useNoAuth } from './useAuth.js'
 export type { UseAuth } from './useAuth.js'
 export { createAuthentication } from './authFactory.js'
+export type { CustomProviderHooks } from './authFactory.js'
 export type { AuthImplementation } from './AuthImplementation.js'
 
 export * from './AuthProvider/AuthProviderState.js'
 export * from './AuthProvider/ServerAuthProvider.js'
+
+export * from './getCurrentUserFromMiddleware.js'
